requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.4
bs4==0.0.1
celery==4.2.0
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
diff-match-patch-python==1.0.2
Django==1.11.5
django-celery-beat==1.1.1
djangorestframework==3.7.7
gunicorn==19.9.0
idna==2.6
kiwisolver==1.0.1
kombu==4.2.1
lxml==4.2.5
matplotlib==3.0.0
numpy==1.15.1
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.3.1
pyparsing==2.2.1
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
wordcloud==1.3.1
